Rename main.py to Making decisions.py
OBJECTIVES
1. become familiar with the input() function;
2. become familiar with comparison operators in Python;
3. become familiar with the concept of conditional execution.

SCENARIO
Spathiphyllum, more commonly known as a peace lily or white sail plant, is one of the most popular indoor houseplants that filter out harmful toxins from the air. Some of the toxins that it neutralizes include benzene, formaldehyde, and ammonia. Imagine that your computer program loves these plants. Whenever it receives an input in the form of the word Spathiphyllum, it involuntarily shouts to the console the following string: "Spathiphyllum is the best plant ever!"

Write a program that utilizes the concept of conditional execution, takes a string as input, and:
1. prints the sentence "Yes - Spathiphyllum is the best plant ever!" to the screen if the inputted string is "Spathiphyllum" (upper-case)
2. prints "No, I want a big Spathiphyllum!" if the inputted string is "spathiphyllum" (lower-case)
3. prints "Spathiphyllum! Not [input]!" otherwise. Note: [input] is the string taken as input.